+++ extensions/chroma/lightID.ts
@@ -1172,6 +1172,7 @@ export const LightIDList: {
    TheRollingStonesEnvironment: {},
    LatticeEnvironment: {},
    DaftPunkEnvironment: {},
+   HipHopEnvironment: {},
    GlassDesertEnvironment: {
       '0': [1, 2, 3, 4, 5, 6],
       '1': [1, 2, 3, 4, 5, 6, 7, 8],