+++ types/beatmap/shared/filename.ts
@@ -5,11 +5,14 @@ export interface IFileInfo {
    filename: string;
 }
 
+// excluded here is official editor way of saving things, otherwise may be saved by 3rd party editor
 export type GenericFilename =
    | `${DifficultyName}${CharacteristicName}.dat`
-   | `${CharacteristicName}${DifficultyName}.dat`
+   | `${Exclude<CharacteristicName, 'Lightshow' | 'Lawless'>}${DifficultyName}.dat`
    | `${DifficultyName}.dat`
-   | `${CharacteristicName}${DifficultyName}.${'beatmap' | 'lightshow'}.dat`
+   | `${Exclude<CharacteristicName, 'Lightshow' | 'Lawless'>}${DifficultyName}.${
+      | 'beatmap'
+      | 'lightshow'}.dat`
    | `${DifficultyName}.${'beatmap' | 'lightshow'}.dat`
    | 'Common.lightshow.dat'
    | 'Lightshow.dat';